+++ src/apps/pendle-v2/pendle-v2.module.ts
@@ -3,18 +3,20 @@ import { AbstractApp } from '~app/app.dynamic-module';
 
 import { PendleV2ContractFactory } from './contracts';
 import { EthereumPendleV2PoolTokenFetcher } from './ethereum/pendle-v2.pool.token-fetcher';
-import { EthereumPendleV2PrincipalTokenTokenFetcher } from './ethereum/pendle-v2.principal-token.token-fetcher';
-import { EthereumPendleV2StandardizedYieldTokenTokenFetcher } from './ethereum/pendle-v2.standardized-yield-token.token-fetcher';
-import { EthereumPendleV2YieldTokenTokenFetcher } from './ethereum/pendle-v2.yield-token.token-fetcher';
+import { EthereumPendleV2PrincipalTokenFetcher } from './ethereum/pendle-v2.principal.token-fetcher';
+import { EthereumPendleV2StandardizedYieldTokenFetcher } from './ethereum/pendle-v2.standardized-yield.token-fetcher';
+import { EthereumPendleV2VotingEscrowContractPositionFetcher } from './ethereum/pendle-v2.voting-escrow.contract-position-fetcher';
+import { EthereumPendleV2YieldTokenFetcher } from './ethereum/pendle-v2.yield.token-fetcher';
 import { PendleV2AppDefinition, PENDLE_V_2_DEFINITION } from './pendle-v2.definition';
 
 @Register.AppModule({
   appId: PENDLE_V_2_DEFINITION.id,
   providers: [
     EthereumPendleV2PoolTokenFetcher,
-    EthereumPendleV2PrincipalTokenTokenFetcher,
-    EthereumPendleV2StandardizedYieldTokenTokenFetcher,
-    EthereumPendleV2YieldTokenTokenFetcher,
+    EthereumPendleV2PrincipalTokenFetcher,
+    EthereumPendleV2StandardizedYieldTokenFetcher,
+    EthereumPendleV2VotingEscrowContractPositionFetcher,
+    EthereumPendleV2YieldTokenFetcher,
     PendleV2AppDefinition,
     PendleV2ContractFactory,
   ],