+++ src/apps/penguin/avalanche/penguin.chef-v1-farm.contract-position-fetcher.ts
@@ -1,49 +1,71 @@
 import { Inject } from '@nestjs/common';
+import { BigNumberish } from 'ethers';
 
 import { APP_TOOLKIT, IAppToolkit } from '~app-toolkit/app-toolkit.interface';
-import { Register } from '~app-toolkit/decorators';
-import { PositionFetcher } from '~position/position-fetcher.interface';
-import { ContractPosition } from '~position/position.interface';
-import { Network } from '~types';
+import { PositionTemplate } from '~app-toolkit/decorators/position-template.decorator';
+import { RewardRateUnit } from '~app-toolkit/helpers/master-chef/master-chef.contract-position-helper';
+import {
+  GetMasterChefDataPropsParams,
+  GetMasterChefTokenBalancesParams,
+  MasterChefTemplateContractPositionFetcher,
+} from '~position/template/master-chef.template.contract-position-fetcher';
 
-import { PenguinContractFactory, PenguinChef } from '../contracts';
-import { PENGUIN_DEFINITION } from '../penguin.definition';
+import { PenguinChef, PenguinContractFactory } from '../contracts';
 
-@Register.ContractPositionFetcher({
-  appId: PENGUIN_DEFINITION.id,
-  groupId: PENGUIN_DEFINITION.groups.chefV1Farm.id,
-  network: Network.AVALANCHE_MAINNET,
-})
-export class AvalanchePenguinChefV1FarmContractPositionFetcher implements PositionFetcher<ContractPosition> {
+@PositionTemplate()
+export class AvalanchePenguinChefV1FarmContractPositionFetcher extends MasterChefTemplateContractPositionFetcher<PenguinChef> {
+  chefAddress = '0x8ac8ed5839ba269be2619ffeb3507bab6275c257';
+  groupLabel = 'Farms';
+  rewardRateUnit = RewardRateUnit.BLOCK;
+
   constructor(
-    @Inject(APP_TOOLKIT) private readonly appToolkit: IAppToolkit,
-    @Inject(PenguinContractFactory) private readonly contractFactory: PenguinContractFactory,
-  ) {}
+    @Inject(APP_TOOLKIT) protected readonly appToolkit: IAppToolkit,
+    @Inject(PenguinContractFactory) protected readonly contractFactory: PenguinContractFactory,
+  ) {
+    super(appToolkit);
+  }
 
-  async getPositions() {
-    return this.appToolkit.helpers.masterChefContractPositionHelper.getContractPositions<PenguinChef>({
-      address: '0x8ac8ed5839ba269be2619ffeb3507bab6275c257',
-      appId: PENGUIN_DEFINITION.id,
-      groupId: PENGUIN_DEFINITION.groups.chefV1Farm.id,
-      network: Network.AVALANCHE_MAINNET,
-      dependencies: [{ appId: 'pangolin', groupIds: ['pool'], network: Network.AVALANCHE_MAINNET }],
-      resolveContract: ({ address, network }) => this.contractFactory.penguinChef({ address, network }),
-      resolvePoolLength: ({ multicall, contract }) => multicall.wrap(contract).poolLength(),
-      resolveDepositTokenAddress: ({ poolIndex, contract, multicall }) =>
-        multicall
-          .wrap(contract)
-          .poolInfo(poolIndex)
-          .then(v => v.lpToken),
-      resolveRewardTokenAddresses: ({ multicall, contract }) => multicall.wrap(contract).pefi(),
-      resolveRewardRate: this.appToolkit.helpers.masterChefDefaultRewardsPerBlockStrategy.build({
-        resolvePoolAllocPoints: async ({ poolIndex, contract, multicall }) =>
-          multicall
-            .wrap(contract)
-            .poolInfo(poolIndex)
-            .then(v => v.allocPoint),
-        resolveTotalAllocPoints: ({ multicall, contract }) => multicall.wrap(contract).totalAllocPoint(),
-        resolveTotalRewardRate: ({ multicall, contract }) => multicall.wrap(contract).pefiPerBlock(),
-      }),
-    });
+  getContract(address: string): PenguinChef {
+    return this.contractFactory.penguinChef({ address, network: this.network });
+  }
+
+  async getPoolLength(contract: PenguinChef): Promise<BigNumberish> {
+    return contract.poolLength();
+  }
+
+  async getStakedTokenAddress(contract: PenguinChef, poolIndex: number): Promise<string> {
+    return contract.poolInfo(poolIndex).then(v => v.lpToken);
+  }
+
+  async getRewardTokenAddress(contract: PenguinChef): Promise<string> {
+    return contract.pefi();
+  }
+
+  async getTotalAllocPoints({ contract }: GetMasterChefDataPropsParams<PenguinChef>): Promise<BigNumberish> {
+    return contract.totalAllocPoint();
+  }
+
+  async getTotalRewardRate({ contract }: GetMasterChefDataPropsParams<PenguinChef>): Promise<BigNumberish> {
+    return contract.pefiPerBlock();
+  }
+
+  async getPoolAllocPoints({ contract, definition }: GetMasterChefDataPropsParams<PenguinChef>): Promise<BigNumberish> {
+    return contract.poolInfo(definition.poolIndex).then(v => v.allocPoint);
+  }
+
+  async getStakedTokenBalance({
+    address,
+    contract,
+    contractPosition,
+  }: GetMasterChefTokenBalancesParams<PenguinChef>): Promise<BigNumberish> {
+    return contract.userInfo(contractPosition.dataProps.poolIndex, address).then(v => v.amount);
+  }
+
+  async getRewardTokenBalance({
+    address,
+    contract,
+    contractPosition,
+  }: GetMasterChefTokenBalancesParams<PenguinChef>): Promise<BigNumberish> {
+    return contract.pendingPEFI(contractPosition.dataProps.poolIndex, address);
   }
 }